+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 ## [0.7.1] - 2025-08-18
 ### Added
  - Optional argument for realization dataset location to override environment variable.
+ - Optional argument for parallel Executor to run_aggregation function.
+
+### Fixed
+ - time zone database on Windows needed for pyarrow.
 
 ## [0.7.0] - 2025-07-28